Added ReplicaError struct to enum variants
