+++ ic-agent/src/agent/http_transport/reqwest_transport.rs
@@ -244,7 +244,7 @@ impl ReqwestTransport {
                 serde_cbor::from_slice(&body);
 
             let agent_error = match cbor_decoded_body {
-                Ok(replica_error) => AgentError::ReplicaErrorV2((replica_error)),
+                Ok(replica_error) => AgentError::ReplicaError(replica_error),
                 Err(cbor_error) => AgentError::InvalidCborData(cbor_error),
             };
