+++ ic-agent/src/agent/mod.rs
@@ -421,13 +421,9 @@ impl Agent {
             .await
             .and_then(|response| match response {
                 replica_api::QueryResponse::Replied { reply } => Ok(reply.arg),
-                replica_api::QueryResponse::Rejected {
-                    reject_code,
-                    reject_message,
-                } => Err(AgentError::ReplicaError {
-                    reject_code,
-                    reject_message,
-                }),
+                replica_api::QueryResponse::Rejected(response) => {
+                    Err(AgentError::ReplicaError(response))
+                }
             })
     }
 
@@ -445,13 +441,9 @@ impl Agent {
             .await
             .and_then(|response| match response {
                 replica_api::QueryResponse::Replied { reply } => Ok(reply.arg),
-                replica_api::QueryResponse::Rejected {
-                    reject_code,
-                    reject_message,
-                } => Err(AgentError::ReplicaError {
-                    reject_code,
-                    reject_message,
-                }),
+                replica_api::QueryResponse::Rejected(response) => {
+                    Err(AgentError::ReplicaError(response))
+                }
             })
     }
 
@@ -542,13 +534,8 @@ impl Agent {
                 reply: Replied::CallReplied(arg),
             } => Ok(PollResult::Completed(arg)),
 
-            RequestStatusResponse::Rejected {
-                reject_code,
-                reject_message,
-            } => Err(AgentError::ReplicaError {
-                reject_code,
-                reject_message,
-            }),
+            RequestStatusResponse::Rejected(response) => Err(AgentError::ReplicaError(response)),
+
             RequestStatusResponse::Done => Err(AgentError::RequestStatusDoneNoReply(String::from(
                 *request_id,
             ))),
